Have negamax return a sentinel -1 in base case
This is safe for two reasons.  First, the AI should never be externally
called to get a move on a finished game, so an invalid response is
reasonable.

Second, the recursive calls to minimax always ignore the returned
index. Only the top-level call cares about the index value, and as per
the first reason above this will always be a "reasonable" value.