Make the game test use new instead of new_game
This way the new_game constructor can not take an injected board.
Currently it's sort of confusing that new game takes a board since it
also calls reset. If only new takes an injected board though, then this
confustion will go away because new_game won't allow you to set a board
so it doesn't matter that it calls reset, and new doesn't call reset.