Change do_nonblocking_turns to only do non-blocking turns
Previously, to make it work with the console it was implicitly forcing
the first turn.  But the key idea here is that if we've just received a
human move request in the web shell then we don't need to block to make
a human move.  But once we've used that up we do.  So we're free to
check solely for moves that will block.