Add a style for "removing" marked spaces
For a dynamic board the easiest way to control display/behavior is to
change a class on the element.  So I made it so that both classes that
cause a mark to be displayed set the radio button for that slot to
"visibility: none" effectively removing it from the page.